Fixes fatal server crash
On Ubuntu (15.10) I get a crash as soon as I add the mdns node to my flow and deploy the flow. After that I have to delete the node out of the flow file, manually, in a text editor to make the flow not crash the server.

The server crashes with the following message:
``
10 Dec 15:16:16 - Error: getaddrinfo -3008
    at errnoException (/usr/lib/node_modules/node-red/node_modules/node-red-node-discovery/node_modules/mdns/lib/resolver_sequence_tasks.js:199:11)
    at getaddrinfo_complete (/usr/lib/node_modules/node-red/node_modules/node-red-node-discovery/node_modules/mdns/lib/resolver_sequence_tasks.js:112:10)
    at GetAddrInfoReqWrap.oncomplete (/usr/lib/node_modules/node-red/node_modules/node-red-node-discovery/node_modules/mdns/lib/resolver_sequence_tasks.js:120:9)
``

I had the same error in my own node.js application that uses mdns. This change should fix it.