Fixed "update" operation
"Update" operation was converting string payloads into an object that contained the "payload" key.  We just want the value of the payload object, not an object containing the key "payload" being inserting into the database.  So instead, the JSON.parse function was used to convert the payload string into an object.