Enable the use of environment variables to define GPIO pin number
The "pin" input field is exposed so that the pin number may be typed in
directly (as opposed to selecting it from the pin table), and also
environment variables may be used to define the pin number. This, in turn,
enables the use of the GPIO pins a subflow template such that multiple
sublfow instances of that template can be running in the same Raspberry Pi
with different pin numbers.