fixed code setting defaults, added floor() and ceil() if result to be integer
I also added a status showing the from and to. This will help people see what is actually being used. i.e you add a random node then attach an inject to send msg.from as 4, you would think the result would be between 4 and 10 but since you did not clear out the 'From' value in the node, the result is from 1 to 10